chore: sync pipeline scripts, code updates, and partial datasets

+++ code/utils.py
@@ -31,7 +31,7 @@ def preprocess(image):
     w, h = image.size
     w, h = map(lambda x: x - x % 32, (w, h))  # resize to integer multiple of 32
     image = image.resize((w, h), resample=Image.LANCZOS)
-    image = np.array(image).astype(np.float16) / 255.0
+    image = np.array(image).astype(np.float32) / 255.0
     image = image.transpose(2,0,1)
     image = torch.from_numpy(image)
     return  2*image - 1.0
